[App] Move fixture into distinct folder
